Add render function to add flexibility and add parse hook
Add a function that will handle different inputs, such as listing views
to update, updating all views, or parsing data.

The parse method is set up to not do anything but allow a class to
inherit the Report class and overload the parse function to be able to
process the data queried from database before it is rendered in a
report.