Fixes #3: Fix bug where ignored views where still rendered
When setting the `ignore` property, it needs to account for both methods
of creating a report; with layout file and with keywords. Previously,
only keywords were handled properly.
Also, by validating the `ignore` property against actual views, it needs
access to the database, so must be instantiated after a `cursor` object.

Remove setup for the private `__views` variable. This was used for the
`views` property, but is no longer needed after `views` rewrite.

Add private `__get_views` function to list all views from database to avoid an
interdependence on the `views` and `ignore` properties. The `ignore` property
used the `views` property to validate entries, and the `views` property used
the `ignore` property to filter out unwanted views.
To avoid this, `__get_views` is used to get all views from the database. This
function can then be used for both properties, without depending on either.